ai-service/requirements.txt
setuptools>=65.0.0
setuptools-rust>=1.3.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.30.0
python-multipart==0.0.20
aiofiles==23.2.1
python-dotenv==1.0.1
openai>=1.52.0
PyPDF2==3.0.1
httpx>=0.27.0
python-docx>=1.1.0
litellm>=1.40.0
openai-whisper>=20231117
moviepy>=1.0.3
